crane: move PCM/WAV utilities from tts.rs to new pcm.rs
These conversions (AudioInfo, pcm_f32_to_i16, encode_wav, save_wav,
load_wav_f32) are direction-agnostic: ASR also needs them, and their
inverses (pcm_i16_to_f32, decode_wav) don't belong under tts.rs.
Pure move, no behavior change.
